Refactor and Format
