+++ client/src/components/App.css
@@ -1,0 +1,48 @@
+.App-logo {
+  height: 40vmin;
+  pointer-events: none;
+}
+
+@media (prefers-reduced-motion: no-preference) {
+  .App-logo {
+    animation: App-logo-spin infinite 20s linear;
+  }
+}
+
+.App {
+  background-color: #282c34;
+  min-height: 100vh;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
+  overflow: hidden;
+}
+
+.App:after {
+  content: "";
+  display: table;
+  clear: both;
+}
+
+.left {
+  text-align: center;
+  float: left;
+  width: 50%;
+}
+
+.right {
+  text-align: center;
+  float: left;
+  width: 50%;
+}
+
+textarea {
+  width: 60%;
+  height: 400px;
+  resize: none;
+  border-radius: 5px;
+}
+
+.App-link {
+  color: #61dafb;
+}
